optimize startup initialization
