+++ src/core/utils/agents_files.ts
@@ -18,12 +18,19 @@ const CHILD_AGENTS_WALK_MAX_DIRS = 10_000;
 
 const DEFAULT_IGNORED_CHILD_DIRS = new Set([
   ".cache",
+  ".cargo",
+  ".config",
+  ".deno",
   ".git",
   ".hg",
   ".jj",
+  ".local",
   ".next",
+  ".npm",
   ".nuxt",
+  ".nvm",
   ".parcel-cache",
+  ".rustup",
   ".svn",
   ".turbo",
   ".venv",
