+++ src/tui/session_chat_app.ts
@@ -23,7 +23,7 @@ export type SessionChatAppOptions = {
   sessionSelection: SessionChatSelection;
   client: TauSdkClient;
   targetLabel: string;
-  terminalAppearance?: ThemeAppearance;
+  terminalAppearance?: ThemeAppearance | Promise<ThemeAppearance>;
   themeId?: string;
   themes?: ThemeDefinition[];
   config?: Config;
@@ -128,12 +128,15 @@ export class SessionChatApp {
 
   static async open(options: SessionChatAppOptions): Promise<SessionChatApp> {
     try {
-      const session = await resolveSession(options.client, options.sessionSelection);
+      const [session, terminalAppearance] = await Promise.all([
+        resolveSession(options.client, options.sessionSelection),
+        options.terminalAppearance,
+      ]);
       const snapshot = await session.snapshot();
       const view = new TuiChatView({
         compactToolUi: true,
         showThinking: false,
-        terminalAppearance: options.terminalAppearance,
+        terminalAppearance,
         themeId: options.themeId,
         themes: options.themes ?? [],
       });
